Fix the usage of None and index() function.
Comparison to None should be performed with 'is' operator instead of '=='. Empty dictionaries are evaluated as False in case of comparison to boolean values.
In case there is no such element in the sequence, .index() method raises a ValueError exception instead of returning None, so comparing the result to None is always False.